implementing (optional) manual length input
admittedly hadn't seen aav's commit when i did this, but making a couple tweaks regardless-

in the previous implementation, manually setting major length is nonfunctional, since nothing in the code toggles SpeedOverride, making it pull major time from the dump 100% of the time. now, when the option for manual setting is ticked, it functions as expected.

also changed the default fallback values to better reflect average update lengths.